Add POST method and file write functionality
The main feature added is a support for HTTP POST method in build_response function. In case of POST request on "/files" path, the content is now saved into a file. This includes the addition of `post_file_response` and `save_file` functions, to handle the POST request and file saving respectively. The constant HTTP response "CREATED" and HTTP method "POST" were added, as well as the extraction of the request method in `parse_request_line` method. Other modifications are related to refactoring and renaming for better code clarity and organization.